backend/requirements.txt
# FastAPI framework and async support
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# STL file processing
numpy-stl==3.2.0

# Payment processing
stripe==7.8.0

# Zoho API integration
requests==2.31.0

# Data validation and models
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Environment variables
python-dotenv==1.0.0

# Email functionality
aiosmtplib==3.0.1
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code quality
ruff==0.1.6
mypy==1.7.1

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4